move remaining torch pruning modifiers to /sparsification/pruning (#571)
* refactor movement and layer pruning modifiers + delete optim/modifier_pruning

* refactor note

* quality

* review fixes

* remove buggy test - to be fixed post release
