+++ src/sparseml/pytorch/optim/__init__.py
@@ -33,7 +33,6 @@ from .modifier_distillation import *
 from .modifier_epoch import *
 from .modifier_lr import *
 from .modifier_params import *
-from .modifier_pruning import *
 from .modifier_regularizer import *
 from .optimizer import *
 from .sensitivity_as import *
