+++ src/sparseml/pytorch/sparsification/pruning/__init__.py
@@ -24,7 +24,9 @@ from .mask_params import *
 from .modifier_pruning_acdc import *
 from .modifier_pruning_base import *
 from .modifier_pruning_constant import *
+from .modifier_pruning_layer import *
 from .modifier_pruning_magnitude import *
 from .modifier_pruning_mfac import *
+from .modifier_pruning_movement import *
 from .modifier_pruning_structured import *
 from .scorer import *
